# sdk/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "whiskd daemon API",
    "version": "17",
    "description": "HTTP/JSON API exposed by the whiskd daemon on loopback. Generated from Go structs; do not edit by hand."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787"
    }
  ],
  "paths": {
...
    "/v1/plugins": {
      "get": {
        "operationId": "listPlugins",
        "summary": "List discovered plugins",
        "tags": [
          "plugins"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PluginStatus"
                  },
                  "type": "array"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/plugins/rescan": {
      "post": {
        "operationId": "rescanPlugins",
        "summary": "Rescan plugin directories",
        "tags": [
          "plugins"
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PluginStatus"
                  },
                  "type": "array"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/plugins/{pluginID}/project-attachment-templates/{templateID}": {
      "post": {
        "operationId": "runPluginProjectAttachmentTemplate",
        "summary": "Run a trusted plugin project attachment template",
        "tags": [
          "plugins"
        ],
        "parameters": [
          {
            "name": "pluginID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunPluginProjectAttachmentTemplateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/plugins/{pluginID}/trust": {
      "post": {
        "operationId": "trustPlugin",
        "summary": "Trust a discovered plugin",
        "tags": [
          "plugins"
        ],
        "parameters": [
          {
            "name": "pluginID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginStatus"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
    "/v1/plugins/{pluginID}/untrust": {
      "post": {
        "operationId": "untrustPlugin",
        "summary": "Untrust a discovered plugin",
        "tags": [
          "plugins"
        ],
        "parameters": [
          {
            "name": "pluginID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PluginStatus"
                }
              }
            },
            "description": "success"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Attachment": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "includeInContext": {
            "type": "boolean"
          },
          "kind": {
            "type": "string"
          },
          "meta": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MetadataValue"
            },
            "type": "object"
          },
          "note": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "scope",
          "createdAt"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "GateConfig": {
        "properties": {
          "blocking": {
            "type": "boolean"
          },
          "command": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "skill": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "blocking"
        ],
        "type": "object"
      },
...
      "MetadataValue": {
        "properties": {
          "bool": {
            "type": "boolean"
          },
          "json": {
            "format": "byte",
            "type": "string"
          },
          "number": {
            "type": "number"
          },
          "string": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "type": "object"
      },
...
      "PluginResolver": {
        "properties": {
          "kinds": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "provider": {
            "type": "string"
          }
        },
        "required": [
          "provider"
        ],
        "type": "object"
      },
      "PluginStatus": {
        "properties": {
          "dir": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "manifestPath": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "projectAttachmentTemplates": {
            "items": {
              "$ref": "#/components/schemas/ProjectAttachmentTemplate"
            },
            "type": "array"
          },
          "resolvers": {
            "items": {
              "$ref": "#/components/schemas/PluginResolver"
            },
            "type": "array"
          },
          "trusted": {
            "type": "boolean"
          },
          "valid": {
            "type": "boolean"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "version",
          "dir",
          "manifestPath",
          "trusted",
          "valid"
        ],
        "type": "object"
      },
      "PluginTemplateField": {
        "properties": {
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "placeholder": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "type"
        ],
        "type": "object"
      },
      "Project": {
        "properties": {
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/Attachment"
            },
            "nullable": true,
            "type": "array"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/MetadataValue"
            },
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "nextWorkItemNumber": {
            "format": "int64",
            "type": "integer"
          },
          "preferences": {
            "$ref": "#/components/schemas/ProjectPreferences"
          },
          "rootDir": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "workflow": {
            "$ref": "#/components/schemas/ProjectWorkflow"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "rootDir",
          "workflow",
          "preferences",
          "attachments",
          "nextWorkItemNumber",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "ProjectAttachmentTemplate": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/PluginTemplateField"
            },
            "type": "array"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "provider",
          "kind"
        ],
        "type": "object"
      },
...
      "ProjectPreferences": {
        "properties": {
          "autoRun": {
            "type": "string"
          },
          "autoWorktree": {
            "type": "boolean"
          },
          "defaultPhaseAgents": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "gates": {
            "items": {
              "$ref": "#/components/schemas/GateConfig"
            },
            "type": "array"
          }
        },
        "required": [
          "autoRun",
          "autoWorktree"
        ],
        "type": "object"
      },
      "ProjectWorkflow": {
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/WorkflowStage"
            },
            "nullable": true,
            "type": "array"
          },
          "templateId": {
            "type": "string"
          },
          "transitionRules": {
            "items": {
              "$ref": "#/components/schemas/TransitionRule"
            },
            "nullable": true,
            "type": "array"
          }
        },
        "required": [
          "id",
          "templateId",
          "name",
          "stages",
          "transitionRules"
        ],
        "type": "object"
      },
...
      "RunPluginProjectAttachmentTemplateRequest": {
        "properties": {
          "projectId": {
            "type": "string"
          },
          "values": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          }
        },
        "required": [
          "projectId"
        ],
        "type": "object"
      },
...
      "TransitionRule": {
        "properties": {
          "fromStageId": {
            "type": "string"
          },
          "requiresApproval": {
            "type": "boolean"
          },
          "requiresChecks": {
            "type": "boolean"
          },
          "requiresNoRunningRuns": {
            "type": "boolean"
          },
          "toStageId": {
            "type": "string"
          }
        },
        "required": [
          "fromStageId",
          "toStageId"
        ],
        "type": "object"
      },
...
      "WorkflowStage": {
        "properties": {
          "defaultPromptTemplateId": {
            "type": "string"
          },
          "defaultRunPreset": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "provisionWorktree": {
            "type": "boolean"
          },
          "wipLimit": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "kind"
        ],
        "type": "object"
      },
...
    }
  }
}